feat: Apple public key generator로 분리
