+++ api/src/main/kotlin/com/nexters/bottles/api/auth/facade/AuthFacade.kt
@@ -1,6 +1,7 @@
 package com.nexters.bottles.api.auth.facade
 
 import com.nexters.bottles.api.auth.component.AppleAuthClientSecretKeyGenerator
+import com.nexters.bottles.api.auth.component.ApplePublicKeyGenerator
 import com.nexters.bottles.api.auth.component.AuthCodeGenerator
 import com.nexters.bottles.api.auth.component.JwtTokenProvider
 import com.nexters.bottles.api.auth.component.NaverSmsEncoder
@@ -34,12 +35,8 @@ import mu.KotlinLogging
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Component
-import java.math.BigInteger
-import java.security.KeyFactory
-import java.security.spec.RSAPublicKeySpec
 import java.time.LocalDate
 import java.time.LocalDateTime
-import java.util.*
 
 
 @Component
@@ -54,6 +51,7 @@ class AuthFacade(
     private val jwtTokenProvider: JwtTokenProvider,
     private val naverSmsEncoder: NaverSmsEncoder,
     private val authCodeGenerator: AuthCodeGenerator,
+    private val applePublicKeyGenerator: ApplePublicKeyGenerator,
     private val appleAuthClientSecretKeyGenerator: AppleAuthClientSecretKeyGenerator,
     private val applicationEventPublisher: ApplicationEventPublisher,
 
@@ -88,21 +86,9 @@ class AuthFacade(
     fun appleSignInUp(appleSignInUpRequest: AppleSignInUpRequest): AppleSignInUpResponse {
         val applePublicKeys = webClientAdapter.sendAppleAuthKeysRequest()
         val tokenHeaders = jwtTokenProvider.parseHeaders(appleSignInUpRequest.code)
-        val applePublicKey = applePublicKeys.keys
-            .findLast { key -> key.kid == tokenHeaders["kid"] && key.alg == tokenHeaders["alg"] }
-            ?: throw IllegalArgumentException("고객센터에 문의해주세요")
-        val nBytes = Base64.getUrlDecoder().decode(applePublicKey.n)
-        val eBytes = Base64.getUrlDecoder().decode(applePublicKey.e)
+        val applePublicKey = applePublicKeyGenerator.generate(applePublicKeys, tokenHeaders["kid"], tokenHeaders["alg"])
+        val appleAccountId = jwtTokenProvider.getAppleTokenClaims(appleSignInUpRequest.code, applePublicKey).subject
 
-        val publicKeySpec = RSAPublicKeySpec(
-            BigInteger(1, nBytes),
-            BigInteger(1, eBytes)
-        )
-
-        val keyFactory = KeyFactory.getInstance(applePublicKey.kty)
-        val publicKey = keyFactory.generatePublic(publicKeySpec)
-
-        val appleAccountId = jwtTokenProvider.getAppleTokenClaims(appleSignInUpRequest.code, publicKey).subject
         val signInUpDto = userService.findAppleUserOrSignUp(appleAccountId)
 
         val userProfile = userProfileService.findUserProfile(signInUpDto.userId)
